Added @Bean annotation so the @Autowired could match and see if it had an object it needed and return it
A bean enables the same instance of that class to be initialised only on starting the app and that same
instance to be used by other classes. @Bean allows something to be dependently injected.  Then added
aws credentials to IntelliJ config and exported them on the terminal to use them when run on intelliJ
or the terminal and noted this on the README and Makefile.